% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run1.R
\name{gbas_model3}
\alias{gbas_model3}
\title{Package gbas_model3}
\usage{
gbas_model3(
  filename,
  pheno_column,
  gene_region,
  anno,
  wt,
  genotype_path,
  phenotype,
  outpath
)
}
\arguments{
\item{filename}{character,vcftools output file}

\item{pheno_column}{numbers}

\item{gene_region}{path, txtfile}

\item{anno}{path,vcffile}

\item{wt}{path,csvfile}

\item{genotype_path}{path}

\item{phenotype}{path,csvfile}

\item{outpath}{outpath}
}
\value{
RDS
}
\description{
Package gbas_model3
}
\examples{
gbas_model3("chr4A",2,"/path/data.gene.list",
"/path/chr4A.annotesplit.filter.vcf",
"/path/wildtype.csv",
"/path/vcftoolsres/",
"/path/phenotype.csv",
"/respath/")

}
