% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gla_model3.R
\name{gla_model3}
\alias{gla_model3}
\title{Package gla_model3}
\usage{
gla_model3(
  filename,
  pheno_column,
  gene_region,
  anno,
  wt,
  genotype_path,
  phenotype,
  outpath
)
}
\arguments{
\item{filename}{character,vcftools output file}

\item{pheno_column}{numbers}

\item{gene_region}{path, txtfile}

\item{anno}{path,txtfile}

\item{wt}{path,csvfile}

\item{genotype_path}{path}

\item{phenotype}{path,csvfile}

\item{outpath}{outpath}
}
\value{
RDS
}
\description{
description of gla_model3
}
\examples{
gla_model3("chr4A",
2,
"/path/data.gene.list",
"/path/chr4A.anno.txt",
"/path/wildtype.csv",
"/path/vcftoolsres/",
"/path/phenotype.csv",
"/respath/")
}
